requirements.txt
Django==1.10.6
djangorestframework==3.6.2
Pygments==2.2.0coreapi==2.3.0
coreapi==2.3.0
